REFACT - make redux structure more inline witht the industry
