+++ src/containers/Qari/index.js
@@ -6,8 +6,8 @@ import Row from 'react-bootstrap/lib/Row';
 import Col from 'react-bootstrap/lib/Col';
 import Button from 'react-bootstrap/lib/Button';
 import Helmet from 'react-helmet';
-import { load, play, next, continuous} from 'redux/modules/audioplayer';
-import { load as loadFiles } from 'redux/modules/files';
+import { load, play, next, continuous} from 'actions/audioplayer';
+import { load as loadFiles } from 'actions/files';
 import zeroPad from 'utils/zeroPad';
 import formatSeconds from 'utils/formatSeconds';
 const styles = require('./style.scss');
