+++ src/reducers/surahs.js
@@ -1,13 +1,12 @@
-import { arrayOf } from 'normalizr';
-import { filesSchema } from '../schemas';
-
-const LOAD = '@@quran/files/LOAD';
-const LOAD_SUCCESS = '@@quran/files/LOAD_SUCCESS';
-const LOAD_FAIL = '@@quran/files/LOAD_FAIL';
+import {
+    LOAD_SUCCESS,
+    LOAD
+} from 'actions/surahs';
 
 const initialState = {
   errored: false,
   loaded: false,
+  current: null,
   entities: {}
 };
 
@@ -26,19 +25,10 @@ export default function reducer(state = initialState, action = {}) {
         errored: false,
         entities: {
           ...state.entities,
-          [action.id]: action.result.entities.files
+          ...action.result.entities.surahs
         }
       };
     default:
       return state;
   }
-}
-
-export function load(id) {
-  return {
-    types: [LOAD, LOAD_SUCCESS, LOAD_FAIL],
-    schema: arrayOf(filesSchema),
-    promise: (client) => client.get(`/qaris/${id}/audio_files/mp3`),
-    id
-  };
 }
