+++ src/redux/create.js
@@ -24,13 +24,12 @@ export default function createStore(history, client, data) {
     finalCreateStore = applyMiddleware(...middleware)(_createStore);
   }
 
-  const reducer = require('./modules/reducer');
+  const reducer = require('../reducers');
   const store = finalCreateStore(reducer, data);
 
-
   if (__DEVELOPMENT__ && module.hot) {
-    module.hot.accept('./modules/reducer', () => {
-      store.replaceReducer(require('./modules/reducer'));
+    module.hot.accept('../reducers', () => {
+      store.replaceReducer(require('../reducers'));
     });
   }
